+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/compatibility/emi/recipe/machine/incubator/AbstractEmiIncubatorRecipe.kt
@@ -1,5 +1,6 @@
 package dev.aaronhowser.mods.geneticsresequenced.compatibility.emi.recipe.machine.incubator
 
+import dev.aaronhowser.mods.geneticsresequenced.util.OtherUtil
 import dev.emi.emi.api.recipe.EmiRecipe
 import dev.emi.emi.api.recipe.EmiRecipeCategory
 import dev.emi.emi.api.recipe.VanillaEmiRecipeCategories
@@ -8,20 +9,16 @@ import dev.emi.emi.api.stack.EmiStack
 import dev.emi.emi.api.widget.WidgetHolder
 import net.minecraft.network.chat.Component
 import net.minecraft.resources.ResourceLocation
-import net.minecraft.world.item.Items
 
 abstract class AbstractEmiIncubatorRecipe : EmiRecipe {
 
     companion object {
-        val BACKGROUND: ResourceLocation =
-            ResourceLocation.withDefaultNamespace("textures/gui/container/brewing_stand.png")
+        val BACKGROUND: ResourceLocation = OtherUtil.modResource("textures/gui/container/incubator_emi.png")
     }
 
     abstract val ingredient: EmiIngredient
     abstract val input: EmiIngredient
     abstract val output: EmiStack
-
-    open val tertiaryItem: EmiStack = EmiStack.of(Items.BLAZE_POWDER)
 
     override fun getCategory(): EmiRecipeCategory {
         return VanillaEmiRecipeCategories.BREWING
@@ -36,7 +33,7 @@ abstract class AbstractEmiIncubatorRecipe : EmiRecipe {
     }
 
     override fun getDisplayWidth(): Int {
-        return 120
+        return 75
     }
 
     override fun getDisplayHeight(): Int {
@@ -46,11 +43,10 @@ abstract class AbstractEmiIncubatorRecipe : EmiRecipe {
     open val tooltips: List<Component> = emptyList()
 
     override fun addWidgets(widgets: WidgetHolder) {
-        widgets.addTexture(BACKGROUND, 0, 0, 103, 61, 16, 14)
-        widgets.addSlot(tertiaryItem, 0, 2).drawBack(false)
-        widgets.addSlot(input, 39, 36).drawBack(false)
-        widgets.addSlot(ingredient, 62, 2).drawBack(false)
-        widgets.addSlot(output, 85, 36).drawBack(false).recipeContext(this)
+        widgets.addTexture(BACKGROUND, 5, 0, 65, 61, 55, 14)
+        widgets.addSlot(input, 5, 36).drawBack(false)
+        widgets.addSlot(ingredient, 28, 2).drawBack(false)
+        widgets.addSlot(output, 51, 36).drawBack(false).recipeContext(this)
 
         widgets.addTooltipText(tooltips, 0, 0, displayWidth, displayHeight)
     }