+++ CHANGELOG.md
@@ -216,3 +216,4 @@
 - The Plasmid Injector EMI recipe page now says in the tooltip that you can have multiple Genes/Antigenes in the same Syringe
 - Fixed an issue in the coloring of Anti-Plasmid tooltips
 - Removed the potion outlines from the Incubator background slots
+- Updated Incubator EMI recipe pages